Schema comparison and on commit validation
