+++ src/odf/dataset-impl/Cargo.toml
@@ -64,8 +64,10 @@ url = { version = "2", default-features = false }
 odf = { workspace = true, features = ["arrow", "lfs", "s3", "testing"] }
 test-utils = { workspace = true }
 
+arrow = { version = "55", default-features = false }
 futures = "0.3"
 mockall = { version = "0.13", default-features = false }
+serde_json = { version = "1", default-features = false }
 tempfile = "3"
 thiserror = { version = "2", default-features = false }
 test-log = { version = "0.2", features = ["trace"] }
